Module rename test
