+++ src/components/Layout/layout.module.scss
@@ -1,0 +1,17 @@
+.container {
+  // background-color: green;
+  margin: 0 auto;
+  max-width: 960px;
+  padding: 0px 1.0875rem 1.45rem;
+  padding-top: 0;
+}
+
+.mainMenuContainer {
+  display: flex;
+  justify-content: space-between;
+  padding: 1rem;
+}
+
+.mainMenu {
+  display: flex;
+}
